Add files via upload
This is one of the projects that I did for my intermediate programming class. For this project, we were using Pygame and had to design the basics of the game for the upcoming final project. We had to build a paddle, a bullet, and a bubble to shoot and appear when you run the main program. The bubble had to move across the top of the screen but not move below the line. The paddle and the bullet had to move along the bottom of the screen together but when you hit the space bar, the bullet would shoot to the top of the screen. For this project, nothing was supposed to happen yet if the bullet hit the bubble.