Add files via upload
This is one of the projects I did for my intermediate programming class. For this project, we used Pygame to create the building blocks of our upcoming final project. We had to create a bubble to get hit by the bullet that moves across the top of the screen but does not move passed the line. We also had to create a paddle and a bullet that moved simultaneously across the bottom of the screen, but if the space bar is hit, then the bullet would shoot across the screen. For this assignment, nothing was to happen yet if the bullet were to hit the bubble.